Fix handling of unsigned requests with trailers
Handling of signed trailers for signature v4 had been added in [1].

  [1]: https://github.com/minio/minio/pull/16484

I've tried to use MinIO with this patch and aws-sdk-cpp [2], but
requests failed with an error:

    Unable to parse ExceptionName: BadRequest Message: malformed chunked encoding

  [2]: https://github.com/aws/aws-sdk-cpp

My requests was unsigned ("x-amz-content-sha256: STREAMING-UNSIGNED-PAYLOAD-TRAILER").

The problem was that it was trying to find AWS "internal chunk"
information (i.e. size in hex), but for such requests there are no such
information, just regular HTTP chunks, so that wrapping
(streaming-v4-unsigned.go) is not required.

After that the problem was that MinIO tries to verify the signature,
however, there is no one, and the request failed with checksum mismatch.

So this patch removes the wrapper for unsigned requests and avoids
checksum checks of unsigned requests, and after this, everything started
working.

P.S. Note that STREAMING-UNSIGNED-PAYLOAD-TRAILER is supported only for
HTTPS [4], so it should be protected from the corruptions anyway.

  [4]: https://github.com/aws/aws-sdk-cpp/blob/0fde1848de338aee2c89ec22887a245a25f21565/src/aws-cpp-sdk-core/source/auth/signer/AWSAuthV4Signer.cpp#L106